backend/requirements.txt
#app
flask == 2.3.2
Flask-Cors == 4.0.0
black == 23.7.0
pymongo == 4.6.0

#Redis
redis == 5.0.1

#socket.io
python-socketio == 5.10.0


#test
flake8 == 6.0.0
pytest == 7.3.1
requests == 2.31.0